+++ .github/workflows/main.yml
@@ -17,4 +17,4 @@ jobs:
           server: shadow.bluewindlab.com
           username: ${{ secrets.ftp_username }}
           password: ${{ secrets.ftp_password }}
-          server-dir: /wp_shadow/
+          server-dir: /pmapi/
