+++ wp-content/themes/bwl-starter-theme/footer.php
@@ -1,0 +1,1 @@
+<?php wp_footer();
