+++ wp-content/themes/bwl-starter-theme/header.php
@@ -1,0 +1,3 @@
+<?php
+
+wp_head();
